Split business-logic (generating the key) from side-effects (grabbing value from env and http/get)
