+++ public/index.php
@@ -1,10 +1,10 @@
 <?php
 
 use Kuria\Error\ErrorHandler;
-use function Siler\Env\env_var;
+use function Siler\array_get_int;
+use function Siler\Env\{env_int, env_var};
 use function Siler\Route\get;
-use function Siler\Twig\init;
-use function Siler\Twig\render;
+use function Siler\Twig\{init, render};
 
 require dirname(__DIR__) . '/vendor/autoload.php';
 
@@ -17,14 +17,44 @@ init(dirname(__DIR__) . '/views');
 
 get('/', function () {
     $length = getKeyLength();
+    list($err, $key) = generateKey($length, getMaxKeyLength());
+
+    if ($err !== null) {
+        stop($err);
+    }
+
     echo render('index.twig', [
-        'key' => generateKey($length),
+        'key' => $key,
          'length'  => $length
     ]);
 });
 
 get('/text', function () {
-    echo  generateKey(
-        getKeyLength()
+    list($err, $key) = generateKey(
+        getKeyLength(),
+        getMaxKeyLength()
     );
+
+    if ($err !== null) {
+        stop($err);
+    }
+
+    echo $key;
 });
+
+function getKeyLength(): int{
+    return array_get_int($_GET, 'length', env_int('KEY_DEFAULT_LENGTH', 64));
+}
+
+function getMaxKeyLength(): int {
+    return env_int('KEY_MAX_LENGTH', 1024);
+}
+
+function stop(string $error, ...$context): void {
+    echo sprintf($error, $context);
+
+    if (function_exists('fastcgi_finish_request')) {
+        fastcgi_finish_request();
+    }
+    exit();
+}
