+++ src/functions.php
@@ -1,33 +1,15 @@
 <?php
 
-use function Siler\array_get_int;
-use function Siler\Env\env_int;
-
-function stop(string $error, ...$context): void {
-    echo sprintf($error, $context);
-
-    if (function_exists('fastcgi_finish_request')) {
-        fastcgi_finish_request();
-    }
-    exit();
-}
-
-function generateKey(int $length) {
-    $max = env_int('KEY_MAX_LENGTH', 1024);
-
+function generateKey(int $length, int $max = 1024): array {
     if ($length <= 0) {
-        stop('Length should be greater than 0');
+        return ['Length should be greater than 0', null];
     }
 
     if ($length > $max) {
-        stop('Length should not be greater than %s', $max);
+        return ["Length should not be greater than $max", null];
     }
 
-    return substr(base64_encode(
+    return [null, substr(base64_encode(
         random_bytes($length)
-    ), 0,  $length);
-}
-
-function getKeyLength(): int{
-    return array_get_int($_GET, 'length', env_int('KEY_DEFAULT_LENGTH', 64));
+    ), 0,  $length)];
 }
